Plugin works!
The UI issue in the 'edit node' (aka Configure in the Node section) is fixed by
loading some initial values on the server side. It seems that on the 'edit'
part, jenkins does not do any Ajax calls. Bad Jenkins.

The plugin can now start, and shutdown vms as needed. We'll probably need to add
some extra logic into it, as for example, number of retries, timeouts, etc, but
this is for the future.

We could also in theory obtain the ip address of the vm that we are running but
for now, this is not entirely necessary since this would require us to install
the ovirt plugin to each vm.

I'll work more on the UI and some minor stuff before releasing a 0.1.0 final
version.